requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1
certifi==2022.12.7
chardet==5.0.0
charset-normalizer==2.0.12; python_version >= '3'
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==1.3.0
django-ckeditor==6.3.2
django-classy-tags==2.0.0
django-cors-headers==3.10.1
django-filter==21.1
django-frontend-forms==0.2.19
django-inlineedit==1.2.2
django-js-asset==2.0.0
django-rest-swagger==2.2.0
django-settings-export==1.2.1
django-sortable-listview==0.43
django-taggit-templatetags2==1.6.1
django-taggit==3.1.0
django-widget-tweaks==1.4.11
django==3.2.18
djangorestframework==3.14.0
gunicorn==20.1.0
httpie==2.6.0
idna==3.4; python_version >= '3'
importlib-metadata==4.8.3; python_version < '3.10'
itypes==1.2.0
jinja2==3.0.3
markdown==3.3.7
markupsafe==2.0.1
mypy-extensions==1.0.0
openapi-codec==1.3.2
pipfile==0.0.2
psycopg2-binary==2.9.6
#psycopg2==2.9.6
pycryptodome==3.17
pygments==2.14.0
pysocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pyyaml==6.0
requests-toolbelt==0.10.1
requests[socks]==2.27.1
simplejson==3.18.4; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
typing-extensions==4.1.1
upgrade-requirements==1.7.0
uritemplate==4.1.1
urllib3==1.26.15
uwsgi==2.0.21
whitenoise==5.3.0
yapps2==2.2.0
zipp==3.6.0
